fix: Switching to a different TOKEN mapping for "From - TO" relationship is not updated in UI if amount is inserted (#160)
- Added network type to address input component
- remove selected resource on destination network change

<!--- Provide a general summary of your changes in the Title above -->

## Description

Resource selection is not updated on UI when a destination network is
changed.

## Related Issue Or Context

https://github.com/sygmaprotocol/sygma-widget/issues/117

When destination network is changed by the user IMO the resource
selected, amount and destination address need to be reset so that they
may update the UI with correct values. I have added functionality that
resets the selected resource, which compels user to select the desired
token again. Furthermore, the address input should be cleared if the
source network is of type `EVM` however, potential destination is of a
different type

Closes: #117 

## Types of changes

- [X] Bug fix (non-breaking change which fixes an issue)